backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "neurovault"
version = "0.1.0"
description = "Type-safe Personal Knowledge Management with AI"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "sqlite-vec>=0.1.0",
    "sentence-transformers>=2.6.0",
    "python-multipart>=0.0.9",
    "aiofiles>=23.0.0",
    "greenlet>=3.0.0",
    "pypdf>=4.0.0",
    "ollama>=0.3.0"
]

[project.optional-dependencies]
dev = [
    "mypy>=1.9.0",
    "pytest>=8.0.0",
    "httpx>=0.27.0"
]

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.setuptools]
packages = ["app", "db", "model"]
